split FriendService and FriendController from UserService/Controller
