+++ backend/src/user/dto/index.ts
@@ -1,4 +1,3 @@
 export * from './createUser.dto';
 export * from './validateUser.dto';
 export * from './updateUserProfile.dto';
-export * from './Friend.dto';
